% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityTable.R
\name{ActivityTable}
\alias{ActivityTable}
\title{Activity Table class definition}
\description{
This R6 class relates to the activity table in the generic report
that documents the person who perfomed the activity and the person who validated
and the respective dates.
Developer note: This class is developed using private and active fields rather
than public fields directly which facilitates more rigid behaviour after the
class has been initialised. If it is deemed that edit the fields after initialisation
is a rare case then we might consider simplifying these classes to use public
fields.
}
\examples{

## ------------------------------------------------
## Method `ActivityTable$new`
## ------------------------------------------------

at <- ActivityTable$new(
 activity = 'Super cool calculation',
 main_author_name = 'David Bowie',
 main_activity_date = '01-Oct-2024',
 qc_author_name = 'Freddie Mercury',
 qc_activity_date = '20-Oct-2024')


## ------------------------------------------------
## Method `ActivityTable$get_table`
## ------------------------------------------------

at <- ActivityTable$new(
 activity = 'Super cool calculation',
 main_author_name = 'David Bowie',
 main_activity_date = '01-Oct-2024',
 qc_author_name = 'Freddie Mercury',
 qc_activity_date = '20-Oct-2024')

at$get_table()

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{activity}}{Active binding for setting the activity}

\item{\code{main_author_name}}{Active binding for setting the main author name}

\item{\code{main_activity_date}}{Active binding for setting the main acitivty date}

\item{\code{qc_author_name}}{Active binding for setting the QC author name}

\item{\code{qc_activity_date}}{Active binding for setting the QC activity date}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActivityTable-new}{\code{ActivityTable$new()}}
\item \href{#method-ActivityTable-get_table}{\code{ActivityTable$get_table()}}
\item \href{#method-ActivityTable-clone}{\code{ActivityTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityTable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityTable$new(
  activity = NA_character_,
  main_author_name = NA_character_,
  main_activity_date = NA_character_,
  qc_author_name = NA_character_,
  qc_activity_date = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity}}{Name of the activity that was performed}

\item{\code{main_author_name}}{Name (Firstname Lastname) of the person who performed the activity}

\item{\code{main_activity_date}}{Date (DD-Mmm-YYYY) when activity was performed.}

\item{\code{qc_author_name}}{Name (Firstname Lastname) of reviewer.}

\item{\code{qc_activity_date}}{Date (DD-Mmm-YYYY) when QC was performed.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{at <- ActivityTable$new(
 activity = 'Super cool calculation',
 main_author_name = 'David Bowie',
 main_activity_date = '01-Oct-2024',
 qc_author_name = 'Freddie Mercury',
 qc_activity_date = '20-Oct-2024')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityTable-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityTable-get_table}{}}}
\subsection{Method \code{get_table()}}{
Show activity table. Takes an object of class \code{\link[rdocx]{ActivityTable}}.
as input and generates Table 1 of the Generic Report Document.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityTable$get_table()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{act}}{An object of class \code{\link[rdocx]{ActivityTable}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Table 1 of the Sample Size Document.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{at <- ActivityTable$new(
 activity = 'Super cool calculation',
 main_author_name = 'David Bowie',
 main_activity_date = '01-Oct-2024',
 qc_author_name = 'Freddie Mercury',
 qc_activity_date = '20-Oct-2024')

at$get_table()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
