% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_figure}
\alias{change_figure}
\title{Change figure in a docx}
\usage{
change_figure(template_doc, full_path, figure)
}
\arguments{
\item{template_doc}{Doc to modify}

\item{full_path}{Path to the folder where the figrue is}

\item{figure}{Figure to be added}
}
\value{
Updated document with new figure
}
\description{
Change figure in a docx
}
\examples{
\dontrun{
change_figure(doc, "/path/to/output", "output.png")
}
}
