% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_table}
\alias{change_table}
\title{Change table in a docx}
\usage{
change_table(template_doc, full_path, table, widths = NA)
}
\arguments{
\item{template_doc}{Doc to modify}

\item{full_path}{Path tot eh folder where the csv is}

\item{table}{Table to be added}

\item{widths}{If there are any pre-specified widths for columns in table}
}
\value{
Updated document with new table
}
\description{
Change table in a docx
}
\examples{
\dontrun{
change_table(doc, "data.csv", widths = NA)
}
}
