% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangelogTable.R
\docType{methods}
\name{ChangelogTable}
\alias{ChangelogTable}
\title{Changelog Table class definition}
\description{
This R6 class controls one row of the Changelog table that documents
the date, version, reason and what changed in the update on the document.
It uses the ChangelogTableRow class to manage the different rows in the table,
and control the checks
}
\examples{

## ------------------------------------------------
## Method `ChangelogTable$add_row`
## ------------------------------------------------

ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first version of the document',
 what_changed = 'Document creation')

changelog_table <- ChangelogTable$new()
changelog_table$add_row(ctr)


## ------------------------------------------------
## Method `ChangelogTable$to_dataframe`
## ------------------------------------------------

ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first versoin of the document',
 what_changed = 'Document creation')

changelog_table <- ChangelogTable$new()
changelog_table$add_row(ctr)
changelog_table$to_dataframe()


## ------------------------------------------------
## Method `ChangelogTable$get_table`
## ------------------------------------------------

ctr_1 <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first versoin of the document',
 what_changed = 'Document creation')

 ctr_2 <- ChangelogTableRow$new(
 date = '15-Oct-2024',
 version = 'Final version',
 why_update = 'Change calculation of param 1',
 what_changed = 'Section 2')

changelog_table <- ChangelogTable$new()
changelog_table$add_row(ctr_1)
changelog_table$add_row(ctr_2)
changelog_table$get_table()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rows}}{List where new rows are added}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChangelogTable-add_row}{\code{ChangelogTable$add_row()}}
\item \href{#method-ChangelogTable-to_dataframe}{\code{ChangelogTable$to_dataframe()}}
\item \href{#method-ChangelogTable-get_table}{\code{ChangelogTable$get_table()}}
\item \href{#method-ChangelogTable-clone}{\code{ChangelogTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangelogTable-add_row"></a>}}
\if{latex}{\out{\hypertarget{method-ChangelogTable-add_row}{}}}
\subsection{Method \code{add_row()}}{
Takes an object of class ChangelogTableRow.
as input and adds it to the list of rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangelogTable$add_row(changelog_table_row_object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{changelog_table_row_object}}{object of class ChangelogTableRow}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first version of the document',
 what_changed = 'Document creation')

changelog_table <- ChangelogTable$new()
changelog_table$add_row(ctr)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangelogTable-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-ChangelogTable-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Create a dataframe from all the rows in the row
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangelogTable$to_dataframe()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first versoin of the document',
 what_changed = 'Document creation')

changelog_table <- ChangelogTable$new()
changelog_table$add_row(ctr)
changelog_table$to_dataframe()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangelogTable-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-ChangelogTable-get_table}{}}}
\subsection{Method \code{get_table()}}{
Generates the changelog table usin to_dataframe()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangelogTable$get_table()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ctr_1 <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first versoin of the document',
 what_changed = 'Document creation')

 ctr_2 <- ChangelogTableRow$new(
 date = '15-Oct-2024',
 version = 'Final version',
 why_update = 'Change calculation of param 1',
 what_changed = 'Section 2')

changelog_table <- ChangelogTable$new()
changelog_table$add_row(ctr_1)
changelog_table$add_row(ctr_2)
changelog_table$get_table()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangelogTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChangelogTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangelogTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
