% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangelogTable.R
\name{ChangelogTableRow}
\alias{ChangelogTableRow}
\title{Class that defines the Changelog Table Row Table in the Generic Report}
\description{
This R6 class controls one row of the Changelog table that documents
the date, version, reason and what changed in the update on the document.
}
\examples{

## ------------------------------------------------
## Method `ChangelogTableRow$new`
## ------------------------------------------------

ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first version of the document',
 what_changed = 'Document creation')

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{date}}{Active binding for setting the date of changes}

\item{\code{version}}{Active binding for setting the version of the document}

\item{\code{why_update}}{Active binding for setting the reason for the update}

\item{\code{what_changed}}{Active binding for setting the section and title impacted}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChangelogTableRow-new}{\code{ChangelogTableRow$new()}}
\item \href{#method-ChangelogTableRow-clone}{\code{ChangelogTableRow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangelogTableRow-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChangelogTableRow-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangelogTableRow$new(
  date = NA_character_,
  version = NA_character_,
  why_update = NA_character_,
  what_changed = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{String. Date (DD-Mmm-YYYY) when modifications were performed}

\item{\code{version}}{String. Document version}

\item{\code{why_update}}{String. Reason for updating the document}

\item{\code{what_changed}}{String. Sections and titles that where impacted with the modifications}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first version of the document',
 what_changed = 'Document creation')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangelogTableRow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChangelogTableRow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangelogTableRow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
