% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_name}
\alias{check_name}
\title{Check name provided}
\usage{
check_name(name_str)
}
\arguments{
\item{name_str}{Name to be checked}
}
\value{
the original date str if matches conditions, FALSE if not
}
\description{
Check to see if name has Firstname Lastname and provide a custom
error to guide users
}
\examples{
\dontrun{
check_name("Harry Styles")
}
}
\keyword{internal}
