% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_null_or_flextable}
\alias{check_null_or_flextable}
\title{Check for NULL or flextable}
\usage{
check_null_or_flextable(variable)
}
\arguments{
\item{variable}{object to check}
}
\value{
the variable if matches conditions, error if not
}
\description{
Checks if the objects passed is either NULL or a flextable
}
\examples{
\dontrun{
ft <- flextable::flextable(head(mtcars))
check_null_or_flextable(ft)
}
}
