% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_string_is_date}
\alias{check_string_is_date}
\title{Check date provided}
\usage{
check_string_is_date(date_str, date_format = "\%d-\%b-\%Y")
}
\arguments{
\item{date_str}{String. Candidate string to check.}

\item{date_format}{String. Expected date format. Default is dd-Mmm-yyyy.}
}
\value{
the original date str if matches conditions, FALSE if not
}
\description{
Check to see if a string is a date format
}
\examples{
\dontrun{
check_string_is_date("01-Oct-2023")
}
}
\keyword{internal}
