% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cursor_reach_list}
\alias{cursor_reach_list}
\title{Cursor reach list}
\usage{
cursor_reach_list(x, keyword)
}
\arguments{
\item{x}{Document to be searched, already read by officer}

\item{keyword}{String. String to be found in the docx document}
}
\value{
List containing the positions where the keyword was found and the
corresponding text associated, as well as the positions where the possible source
path could be and its associated source path text
}
\description{
This function read a word document with xml2, looks for a pattern
in the text (i.e Figure/tables captions)/ Return the positions and text for this
search, and also possible sources path underneath the placeholder figure/tables
}
\examples{
\dontrun{
cursor_reach_list(doc, "example caption")
}
}
\keyword{internal}
