% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_caption}
\alias{escape_caption}
\title{Escape special characters in a string for regular expresion}
\usage{
escape_caption(caption)
}
\arguments{
\item{caption}{Caption/Title of a figure}
}
\value{
escaped caption
}
\description{
This function escapes special characters in a given caption
string using the gsub function. Special characters can cause issues in certain
contexts, such as regular expressions or when processing text data. By escaping
these characters, the resulting string can be safely used in various applications.
}
\details{
The function uses the gsub function to replace special characters in the
input caption with their escaped versions. The regular expression pattern includes
a character class specifying the set of characters to be escaped.
The replacement string contains four backslashes followed by the matched special
character to ensure proper escaping.
}
\examples{
\dontrun{
caption <- "This is a figure title (an a parenthesis use!)"
escape_caption(caption)
}
}
\keyword{internal}
