% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_figure_table_captions}
\alias{extract_figure_table_captions}
\title{Extract figure and table captions}
\usage{
extract_figure_table_captions(doc_path)
}
\arguments{
\item{doc_path}{String. Path to the docx document}
}
\value{
List figure_table_captions_sources contaning the figure_table_captions
and its corresponding sources paths
}
\description{
given a word document, extract all the caption for figures and
tables using 'cursor_reach_list()'. If the docx has a ToC, 'cursor_reach_list()'
will return captions from both the body of the text and the ToC. For that,
'extract_figure_table_captions()' cleans the captions based on keywords
contained internally in the docx. ToC captions (keyword PAGEREF) but the first
caption one is always missed by the searching function. Body captions
(keyword STYLEREF or nothing) are always found correctly.
Therefore, if STYLEREF is present we filter for those. If there is no STYLEREF
but there is PAGEREF, we take the ones that are not PAGEREF. If not STYLEREF
or PAGEREF, no need to clean anything.
}
\examples{
\dontrun{
extract_figure_table_captions(system.file("use_cases/02_automated_reporting",
                                         "Automated_Reporting_Example.docx",
                                         package="rdocx"))
}
}
