% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocx_global.R
\name{generic_report_template}
\alias{generic_report_template}
\title{Generic Report Template location}
\usage{
generic_report_template()
}
\description{
Function which returns the location of the generic report template
}
