% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_fig_table_part}
\alias{remove_fig_table_part}
\title{Remove Figure/Table part of caption}
\usage{
remove_fig_table_part(caption)
}
\arguments{
\item{caption}{String. Caption that need to be modified}
}
\value{
updated caption without Figure or Table number
}
\description{
Used to remove the first part fo the captions. It will clean
normal captions like Figure/Table followed by its numbering (i.e. Figure 1-1,
Table 2-1), and also more complex patterns created by using Novstyle such as
"Table  STYLEREF 1 \\s 2 SEQ Table \\* ARABIC \\s 1 1. Summary of Sample Statistics"
}
\examples{
\dontrun{
example_caption <- 
"Table  STYLEREF 1 \\\\s 2 SEQ Table \\\\* ARABIC \\\\s 1 1. Summary of Sample Statistics"
updated_example_caption <- remove_fig_table_part(example_caption)
expected_output <- "Summary of Sample Statistics"
}
}
\keyword{internal}
