% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_text}
\alias{replace_text}
\title{Replace text with officer}
\usage{
replace_text(doc, old_value, new_value)
}
\arguments{
\item{doc}{Docx document to be modified}

\item{old_value}{Word to search in the doc}

\item{new_value}{Replacement for the old_value}
}
\description{
Search for a value in a document and replicate it with the new one
}
\examples{
\dontrun{
replace_text("dummy_doc.docx", "title", "new title")
}
}
