% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_render.R
\name{rmd_render}
\alias{rmd_render}
\title{Rmd Sample Size renderer}
\usage{
rmd_render(
  rmd_filename,
  output_path = dirname(rmd_filename),
  version = 0,
  reference_docx = generic_report_template()
)
}
\arguments{
\item{rmd_filename}{String. Name of the Rmd file to be rendered.}

\item{output_path}{String. Set the path where your generated Word file will be written to.
The default path will the path of the Rmd file.}

\item{version}{Integer. Version number of the created Word file, e.g., 0, 1, 12.
Will be modified to follow version numbering guidelines, e.g., _v00, _v01, v_12.}

\item{reference_docx}{String. Template document in .docx format providing
styling and structure information for rendering your docx report.
This is configured to be the generic report template using the function
\code{generic_report_template()}.}
}
\value{
Word docx file rendered from the Rmd located in \code{output_path}
}
\description{
Given the filename of a generic report Rmd template, it renders it.
It is a wrapper that hides the process of rendering, starting by getting the
title and body separately, getting the study code and compound name, and merging
both files with the correct output name.
}
\examples{
\dontrun{
rmd_render(system.file("use_cases/01_generic_report",
                       "generic_report_template.Rmd",
                       package="rdocx"))
}
}
