% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_updated_document}
\alias{save_updated_document}
\title{Save updated document}
\usage{
save_updated_document(document, doc_final_path)
}
\arguments{
\item{document}{Officer doc to be saved}

\item{doc_final_path}{Location to save}
}
\description{
Save updated document
}
\examples{
\dontrun{
save_updated_document(doc, "./new_doc.docx")
}
}
\keyword{internal}
