% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_for_duplicates}
\alias{search_for_duplicates}
\title{Search for duplicates}
\usage{
search_for_duplicates(yaml_file)
}
\arguments{
\item{yaml_file}{List. Already read yaml file to be checked for duplicates}
}
\description{
Look for duplicates in captions titles and raise a warning if
the occurrence parameter is not defined.
}
\keyword{internal}
