% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturesTable.R
\docType{methods}
\name{SignaturesTable}
\alias{SignaturesTable}
\title{Signatures Table class definition}
\description{
This R6 class controls one row of the Signatures table that documents
the date, version, reason and what changed in the update on the document.
It uses the ChangelogTableRow class to manage the different rows in the table,
and control the checks
}
\examples{

## ------------------------------------------------
## Method `SignaturesTable$add_row`
## ------------------------------------------------

str <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')


signatures_table <- SignaturesTable$new()
signatures_table$add_row(str)


## ------------------------------------------------
## Method `SignaturesTable$to_dataframe`
## ------------------------------------------------

str <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')


signatures_table <- SignaturesTable$new()
signatures_table$add_row(str)
signatures_table$to_dataframe()


## ------------------------------------------------
## Method `SignaturesTable$get_table`
## ------------------------------------------------

str_1 <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')

str_2 <- SignaturesTableRow$new(
 name = 'Alex Turner',
 department = 'Product Development',
 date = '20-Feb-2024')

signatures_table <- SignaturesTable$new()
signatures_table$add_row(str_1)
signatures_table$add_row(str_2)
signatures_table$get_table()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rows}}{List where new rows are added}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignaturesTable-add_row}{\code{SignaturesTable$add_row()}}
\item \href{#method-SignaturesTable-to_dataframe}{\code{SignaturesTable$to_dataframe()}}
\item \href{#method-SignaturesTable-get_table}{\code{SignaturesTable$get_table()}}
\item \href{#method-SignaturesTable-clone}{\code{SignaturesTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignaturesTable-add_row"></a>}}
\if{latex}{\out{\hypertarget{method-SignaturesTable-add_row}{}}}
\subsection{Method \code{add_row()}}{
Takes an object of class SignaturesTableRow.
as input and adds it to the list of rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignaturesTable$add_row(signatures_table_row_object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signatures_table_row_object}}{object of class SignaturesTableRow}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{str <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')


signatures_table <- SignaturesTable$new()
signatures_table$add_row(str)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignaturesTable-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-SignaturesTable-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Create a dataframe from all the rows in the row
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignaturesTable$to_dataframe()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{str <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')


signatures_table <- SignaturesTable$new()
signatures_table$add_row(str)
signatures_table$to_dataframe()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignaturesTable-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-SignaturesTable-get_table}{}}}
\subsection{Method \code{get_table()}}{
Generates the signature table usin to_dataframe()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignaturesTable$get_table()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{str_1 <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')

str_2 <- SignaturesTableRow$new(
 name = 'Alex Turner',
 department = 'Product Development',
 date = '20-Feb-2024')

signatures_table <- SignaturesTable$new()
signatures_table$add_row(str_1)
signatures_table$add_row(str_2)
signatures_table$get_table()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignaturesTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SignaturesTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignaturesTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
