% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturesTable.R
\name{SignaturesTableRow}
\alias{SignaturesTableRow}
\title{Signatures Table Row class definition}
\description{
This R6 class controls one row of the Signatures table that documents
the date, version, reason and what changed in the update on the document.
}
\examples{

## ------------------------------------------------
## Method `SignaturesTableRow$new`
## ------------------------------------------------

str <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Active binding for setting the name of the person who will sign}

\item{\code{department}}{Active binding for setting the department of the person who signs}

\item{\code{date}}{Active binding for setting the date of sign}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignaturesTableRow-new}{\code{SignaturesTableRow$new()}}
\item \href{#method-SignaturesTableRow-clone}{\code{SignaturesTableRow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignaturesTableRow-new"></a>}}
\if{latex}{\out{\hypertarget{method-SignaturesTableRow-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignaturesTableRow$new(
  name = NA_character_,
  department = NA_character_,
  date = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String. Name of the person who will sign the document}

\item{\code{department}}{String. Department of the person who will sign the document}

\item{\code{date}}{String. Date (DD-Mmm-YYYY) when the document will be signed}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{str <- SignaturesTableRow$new(
 name = 'Harry Styles',
 department = 'eCompliance',
 date = '15-Feb-2024')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignaturesTableRow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SignaturesTableRow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignaturesTableRow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
