% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_styling}
\alias{table_styling}
\title{Add Generic Report Style to a flextable}
\usage{
table_styling(flex_table, widths = NA)
}
\arguments{
\item{flex_table}{A flextable to modify style}

\item{widths}{If there is a specified width for the columns}
}
\value{
Flextable with Generic Report Style
}
\description{
Add Generic Report Style to a flextable
}
\examples{
\dontrun{
table_styling(table, widths = NA)
}
}
