% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TitlePage.R
\name{TitlePage}
\alias{TitlePage}
\alias{get_title_page}
\title{Title page class definition}
\description{
This R6 class relates to the Title Page in a report that adds all
the required fields of the Generic Report Title Page
Developer note: This class is developed using private and active fields rather
than public fields directly which facilitates more rigid behaviour after the
class has been initialised. If it is deemed that edit the fields after initialisation
is a rare case then we might consider simplifying these classes to use public
fields.
}
\examples{

## ------------------------------------------------
## Method `TitlePage$new`
## ------------------------------------------------

\dontrun{
tp = TitlePage$new(
  report_title = "Super cool document",
  department = "Cool department",
  study_title = "Compute super cool things",
  study_number = "OWNDJQW9923",
  status = "Draft",
  version  = "1",
  date = "01-Feb-2024",
  bus_class = "Very confidential")
  }

## ------------------------------------------------
## Method `TitlePage$get_title_page`
## ------------------------------------------------

\dontrun{
tp = TitlePage$new(
  report_title= "Super cool document",
  department = "Cool department",
  study_title = "Compute super cool things",
  study_number = "OWNDJQW9923",
  status = "Draft",
  version  = "1",
  date = "01-Feb-2024",
  bus_class = "Very confidential")
tp$get_title_page()
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{report_title}}{Active binding for setting the report title}

\item{\code{department}}{Active binding for setting the department}

\item{\code{study_title}}{Active binding for setting the study title}

\item{\code{study_number}}{Active binding for setting the study number}

\item{\code{status}}{Active binding for setting the document status}

\item{\code{version}}{Active binding for setting the document version}

\item{\code{date}}{Active binding for setting the date}

\item{\code{bus_class}}{Active binding for setting the business classification}

\item{\code{changelog_table}}{Active binding for setting the changelog table}

\item{\code{signatures_table}}{Active binding for setting the signatures table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TitlePage-new}{\code{TitlePage$new()}}
\item \href{#method-TitlePage-get_title_page}{\code{TitlePage$get_title_page()}}
\item \href{#method-TitlePage-clone}{\code{TitlePage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TitlePage-new"></a>}}
\if{latex}{\out{\hypertarget{method-TitlePage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TitlePage$new(
  report_title = NA_character_,
  department = NA_character_,
  study_title = NA_character_,
  study_number = NA_character_,
  status = c("Draft", "Final"),
  version = NA_character_,
  date = NA_character_,
  bus_class = NA_character_,
  changelog_table = NULL,
  signatures_table = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_title}}{String. Report title.}

\item{\code{department}}{String. Department.}

\item{\code{study_title}}{String. Study title of the document}

\item{\code{study_number}}{String. Study number of the document}

\item{\code{status}}{String. Status of the document, with possible values \code{Draft} or \code{Final}}

\item{\code{version}}{String. Version of the document.}

\item{\code{date}}{String. Date (DD-Mmm-YYYY) of the final document release.}

\item{\code{bus_class}}{String. Business Classification.}

\item{\code{changelog_table}}{Flextable. Table to be added after the title page (Optional)}

\item{\code{signatures_table}}{Flextable. Table to be added after the title page (Optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
tp = TitlePage$new(
  report_title = "Super cool document",
  department = "Cool department",
  study_title = "Compute super cool things",
  study_number = "OWNDJQW9923",
  status = "Draft",
  version  = "1",
  date = "01-Feb-2024",
  bus_class = "Very confidential")
  }
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TitlePage-get_title_page"></a>}}
\if{latex}{\out{\hypertarget{method-TitlePage-get_title_page}{}}}
\subsection{Method \code{get_title_page()}}{
Generate title page. Takes an object of class \code{\link[rdocx]{TitlePage}}.
as input and generates the title page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TitlePage$get_title_page(
  reference_docx = generic_report_template(),
  output_path = "."
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference_docx}}{String. Template document in .docx format providing
styling and structure information for rendering your docx report.
This is configured to be Generic Report Template using the function
\code{generic_report_template()}.}

\item{\code{output_path}}{String. Path where the title page in docx format should
be written. Default is the current working directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A title page in Word format ('_generic_report_title.docx').
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
tp = TitlePage$new(
  report_title= "Super cool document",
  department = "Cool department",
  study_title = "Compute super cool things",
  study_number = "OWNDJQW9923",
  status = "Draft",
  version  = "1",
  date = "01-Feb-2024",
  bus_class = "Very confidential")
tp$get_title_page()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TitlePage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TitlePage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TitlePage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
