% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{version_number}
\alias{version_number}
\title{Version number}
\usage{
version_number(version)
}
\arguments{
\item{version}{Number. The current version of the document}
}
\value{
formatted version "_v00", "_v02", _v13"
}
\description{
Version number
}
\examples{
\dontrun{
version_number(5)
}
}
